wally/requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
rich>=13.7.1
